pyproject.toml
[tool.poetry]
name = "netlify-builds"
version = "0.1.0"
description = "A command line utility to check build usage across multiple Netlify accounts"
authors = ["Bruno Alla <alla.brunoo@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/browniebroke/netlify-builds"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "netlify_builds" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/browniebroke/netlify-builds/issues"
"Changelog" = "https://netlify-builds.readthedocs.io/changelog.html"

[tool.poetry.scripts]
netlify-builds = "netlify_builds.main:main"

[tool.poetry.dependencies]
python = "^3.6"
httpx = "^0.16.1"
python-dateutil = "^2.8.1"
rich = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
pytest = "^6.1.2"
pytest-cov = "^2.10.1"
pyupgrade = {version = "^2.7.4", python = "^3.6.1"}
tox = "^3.20.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
branch = "main"
version_variable = [
    "netlify_builds/__init__.py:__version__",
    "pyproject.toml:version",
]
build_command = "pip install poetry && poetry build"

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=netlify_builds"

[tool.isort]
profile = "black"
known_first_party = ["netlify_builds", "tests"]
